Replace autotools with CMake
Replace the autotools build system with CMake. I was going to try to
preserve both, but ran into some things that would have been a headache;
particularly, use of generic preprocessor defines that, in case of a
static build, would have leaked into consumers and potentially caused
problems.

Currently, this does not include coverage or doxygen generation,
although it's unclear that the old autotools supported doxygen as part
of the build system anyway (the instructions are to just run 'doxygen'
in the source tree). This should also Just Work on Windows/MSVC without
needing to maintain separate build files, as was apparently being done.